+++ addon/views/ivy-sortable.js
@@ -5,7 +5,19 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
 
   tagName: 'ul',
 
+  arrayDidChangeAfterElementInserted: function() {
+    Ember.run.scheduleOnce('afterRender', this, this._refreshSortable);
+  },
+
+  arrayWillChangeAfterElementInserted: function() {
+  },
+
   destroySortable: Ember.on('willDestroyElement', function() {
+    this._contentWillChangeAfterElementInserted();
+
+    Ember.removeBeforeObserver(this, 'content', this, this._contentWillChangeAfterElementInserted);
+    this.removeObserver('content', this, this._contentDidChangeAfterElementInserted);
+
     this.$().sortable('destroy');
   }),
 
@@ -24,6 +36,11 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
       'opacity', 'placeholder', 'revert', 'scroll', 'scrollSensitivity',
       'scrollSpeed', 'tolerance', 'zIndex'
     ], this._bindSortableOption, this);
+
+    Ember.addBeforeObserver(this, 'content', this, this._contentWillChangeAfterElementInserted);
+    this.addObserver('content', this, this._contentDidChangeAfterElementInserted);
+
+    this._contentDidChangeAfterElementInserted();
   }),
 
   move: function(oldIndex, newIndex) {
@@ -101,6 +118,34 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
     });
   },
 
+  _contentDidChangeAfterElementInserted: function() {
+    var content = this.get('content');
+
+    if (content) {
+      content.addArrayObserver(this, {
+        didChange: 'arrayDidChangeAfterElementInserted',
+        willChange: 'arrayWillChangeAfterElementInserted'
+      });
+    }
+
+    var len = content ? Ember.get(content, 'length') : 0;
+    this.arrayDidChangeAfterElementInserted(content, 0, null, len);
+  },
+
+  _contentWillChangeAfterElementInserted: function() {
+    var content = this.get('content');
+
+    if (content) {
+      content.removeArrayObserver(this, {
+        didChange: 'arrayDidChangeAfterElementInserted',
+        willChange: 'arrayWillChangeAfterElementInserted'
+      });
+    }
+
+    var len = content ? Ember.get(content, 'length') : 0;
+    this.arrayWillChangeAfterElementInserted(content, 0, len);
+  },
+
   _disableArrayObservers: function(content, callback) {
     content.removeArrayObserver(this);
     try {
@@ -112,5 +157,9 @@ export default Ember.CollectionView.extend(Ember.TargetActionSupport, {
 
   _optionDidChange: function(sender, key) {
     this.$().sortable('option', key, this.get(key));
+  },
+
+  _refreshSortable: function() {
+    this.$().sortable('refresh');
   }
 });